+++ hybrid_core/hybrid_core/example/lib/view/os_view.dart
@@ -2,11 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:hybrid_core/hybrid_core.dart';
 
 class OSView extends StatelessWidget {
-  const OSView({super.key});
+  final OS os;
+
+  OSView({super.key}) : os = OS();
 
   @override
   Widget build(BuildContext context) {
-    final os = OS();
+    final os = this.os;
     final String version;
     final bool atLeastVersion;
     if (os is Android) {
@@ -20,6 +22,9 @@ class OSView extends StatelessWidget {
       version = '${os.version}';
       final version14_0 = DarwinVersion.number(14.0);
       atLeastVersion = os.atLeastVersion(version14_0);
+    } else if (os is Windows) {
+      version = '${os.version}';
+      atLeastVersion = os.isWindows10OrGreater;
     } else {
       throw TypeError();
     }
@@ -43,7 +48,7 @@ class OSView extends StatelessWidget {
               height: 20.0,
             ),
             Text(
-              'At least API 33+ or iOS 17.0+ or macOS 14.0+',
+              'At least API 33+ or iOS 17.0+ or macOS 14.0+ or Windows 10+',
               style: Theme.of(context).textTheme.bodyMedium,
             ),
             Text(