+++ hybrid_core/hybrid_core_android/lib/hybrid_core_android.dart
@@ -1,11 +1,2 @@
-import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
-
-import 'src/jni.dart';
-import 'src/os.dart';
-
-abstract class HybridCoreAndroid {
-  static void registerWith() {
-    Jni.initDLApi();
-    OSPlatform.instance = OSPlatformImpl();
-  }
-}
+export 'src/hybrid_core_android.dart';
+export 'src/os.dart';