+++ hybrid_core/hybrid_core_linux/lib/hybrid_core_linux.dart
@@ -1,3 +1,1 @@
-abstract class HybridCoreLinux {
-  static void registerWith() {}
-}
+export 'src/hybrid_core_linux.dart';