+++ hybrid_core/hybrid_core_windows/lib/src/ffi.dart
@@ -1,0 +1,6 @@
+export 'dart:ffi';
+
+export 'package:ffi/ffi.dart';
+
+export 'ffi.g.dart';
+export 'macros.dart';