+++ hybrid_core/hybrid_core_windows/lib/src/hybrid_core_windows.dart
@@ -1,0 +1,9 @@
+import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
+
+import 'os.dart';
+
+abstract class HybridCoreWindows {
+  static void registerWith() {
+    OSPlatform.instance = WindowsPlatform();
+  }
+}