+++ hybrid_core/hybrid_core_windows/lib/src/macros.dart
@@ -1,0 +1,4 @@
+// ignore_for_file: type=lint
+
+int LOBYTE(int w) => w & 0xff;
+int HIBYTE(int w) => (w >> 8) & 0xff;