+++ hybrid_core/hybrid_core_platform_interface/CHANGELOG.md
@@ -7,3 +7,13 @@
 * Move `AndroidOS` to `Android`.
 * Move `DarwinOS` to `Darwin`.
 * Move `DarwinOSVersion` to `DarwinVersion`.
+
+## 0.2.0
+
+* Add `OS#renderWidgetToMemory` method.
+* Remove `OSPlatform#os` property.
+* Remove `Android` class.
+* Remove `Darwin` class.
+* Remove `iOS` class.
+* Remove `macOS` class.
+* Remove `DarwinVersion` class.