+++ hybrid_core/hybrid_core_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
 * Add support for Windows.
+
+## 0.2.0
+
+* Implement `OSPlatform` interface.